refactor(server): finish mcp_server.py -> server.py transition
IG's Step 5 retrofit renamed create_server() -> make_server() but left
the file at src/image_generation_mcp/mcp_server.py and kept a
'create_server = make_server' compatibility alias.  That half-state
caused the weekly copier-update cron's generated tests/test_smoke.py
to fail: it imports 'image_generation_mcp.server.make_server' (the
template's canonical name) which didn't exist.

Finishes the transition:

- git mv mcp_server.py -> server.py
- all import/patch sites migrated from image_generation_mcp.mcp_server
  to image_generation_mcp.server (15 files, all via sed)
- all create_server call sites migrated to make_server
- drop the 'create_server = make_server' alias
- pyproject.toml per-file-ignores updated (mcp_server.py -> server.py);
  TC002 added for starlette/EventStore imports that are intentionally
  runtime (not TYPE_CHECKING) for make_server's conditional custom_route
  wiring
- tests/test_smoke.py restored to the template's canonical version
  (no longer needs the no-op stub workaround)
- CLAUDE.md project-structure table + Key Patterns line updated

Gates green: ruff clean, ruff format clean, mypy clean, 807 tests pass.
