+++ tests/test_resources_as_tools.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING
 import pytest
 from fastmcp import Client
 
-from image_generation_mcp.mcp_server import create_server
+from image_generation_mcp.server import make_server
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -18,23 +18,23 @@ class TestResourcesAsToolsRegistration:
     """Verify ResourcesAsTools generates list_resources and read_resource tools."""
 
     async def test_list_resources_tool_exists(self) -> None:
-        server = create_server()
+        server = make_server()
         tool_names = [t.name for t in await server.list_tools()]
         assert "list_resources" in tool_names
 
     async def test_read_resource_tool_exists(self) -> None:
-        server = create_server()
+        server = make_server()
         tool_names = [t.name for t in await server.list_tools()]
         assert "read_resource" in tool_names
 
     async def test_original_tools_still_present(self) -> None:
-        server = create_server()
+        server = make_server()
         tool_names = [t.name for t in await server.list_tools()]
         assert "list_providers" in tool_names
 
     async def test_hand_written_get_image_removed(self) -> None:
         """get_image and list_images are no longer registered as tools."""
-        server = create_server()
+        server = make_server()
         tool_names = [t.name for t in await server.list_tools()]
         assert "get_image" not in tool_names
         assert "list_images" not in tool_names
@@ -44,7 +44,7 @@ class TestResourcesAsToolsReadOnly:
     """Verify transform works in both read-only and read-write modes."""
 
     async def test_read_only_has_resource_tools(self) -> None:
-        server = create_server()
+        server = make_server()
         tool_names = [t.name for t in await server.list_tools()]
         assert "list_resources" in tool_names
         assert "read_resource" in tool_names
@@ -53,7 +53,7 @@ class TestResourcesAsToolsReadOnly:
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
         monkeypatch.setenv("IMAGE_GENERATION_MCP_READ_ONLY", "false")
-        server = create_server()
+        server = make_server()
         tool_names = [t.name for t in await server.list_tools()]
         assert "list_resources" in tool_names
         assert "read_resource" in tool_names
@@ -62,7 +62,7 @@ class TestResourcesAsToolsReadOnly:
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
         monkeypatch.setenv("IMAGE_GENERATION_MCP_READ_ONLY", "false")
-        server = create_server()
+        server = make_server()
         tool_names = [t.name for t in await server.list_tools()]
         assert "generate_image" in tool_names
         assert "list_resources" in tool_names
@@ -82,7 +82,7 @@ class TestResourcesAsToolsEndToEnd:
         scratch dir so the result is a deterministic empty JSON array.
         """
         monkeypatch.setenv("IMAGE_GENERATION_MCP_SCRATCH_DIR", str(tmp_path))
-        server = create_server()
+        server = make_server()
         async with Client(server) as client:
             result = await client.call_tool("read_resource", {"uri": "image://list"})
